ENPM611-Team_1_Poetry-Issue-Analyzer-main/ENPM611-Team_1_Poetry-Issue-Analyzer-main/requirements.txt
python-dateutil
pandas
matplotlib
tabulate